move parse-server-status.ts
